requirements.txt
tensorflow
pillow
matplotlib
fastapi
uvicorn[standard]
kafka-python
python-multipart
hdfs